Avoid deleting unexistent files

+++ utils/cleanup.go
@@ -7,9 +7,11 @@ import (
 
 func CleanUpFiles(filePaths []string) {
 	for _, filePath := range filePaths {
-		err := os.Remove(filePath)
-		if err != nil {
-			log.Fatal(err)
+		if _, err := os.Stat(filePath); err == nil {
+			err := os.Remove(filePath)
+			if err != nil {
+				log.Fatal(err)
+			}
 		}
 	}
 }
